Bug 1466917: Explicitly set link up PUBLIC_INTERFACE after br-ex add
when using ODL_L3 in lithium, it is now possible to have north/south
support via floating ips. to support that, networking-odl will pay
attention to the public_interface attribute and add it as a port [1]
of the external bridge.

Besides doing that, it should also "link ip" the public interface,
just as how that is done in other port add implementations [2].

[1]: https://review.openstack.org/#/c/183032/
[2]: https://github.com/openstack-dev/devstack/blob/master/lib/neutron-legacy#L558

Change-Id: I52a7b5ade8f45b79cfc67838bd4e0db2944adb0b