Add ODL_OVS_MANAGERS to support clustering
Introducing ODL_OVS_MANAGERS, an optional variable.
In non-clustering cases, this is normally the same as ODL_MGR_IP.
However, for HA deployments the southbound portion to ODL is expected
to use the ip addresses of the ODL instances instead of a single vip.
That enables OVS to simultaneouly connect to more than one ODL instance.

Example of expected format:
ODL_OVS_MANAGERS=1.1.1.1,2.2.2.2,3.3.3.3